xmhf: xmhf-hic: start digesting SMP bringup code in 32-bit...

+++ xmhf-hic/include/xmhf-hic.h
@@ -328,7 +328,7 @@ void xmhfhic_arch_setup_slab_mem_page_tables(void);
 void xmhfhic_arch_switch_to_smp(void);
 void xmhfhic_arch_setup_base_cpu_data_structures(void);
 void xmhf_hic_arch_setup_cpu_state(u64 cpuid);
-void xmhfhic_smp_entry(u64 cpuid);
+void xmhfhic_smp_entry(u32 cpuid);
 
 
 extern void xmhfhic_arch_relinquish_control_to_init_slab(u64 cpuid, u64 entrystub, u64 mempgtbl_cr3, u64 slabtos);
@@ -380,7 +380,8 @@ extern __xmhfhic_safestack_element_t __xmhfhic_safestack[MAX_PLATFORM_CPUS][512]
 extern __attribute__(( aligned(16) )) u64 __xmhfhic_x86vmx_gdt_start[];     //ro
 extern __attribute__(( aligned(16) )) arch_x86_gdtdesc_t __xmhfhic_x86vmx_gdt;  //ro
 extern __attribute__(( aligned(4096) )) u8 __xmhfhic_x86vmx_tss[MAX_PLATFORM_CPUS][PAGE_SIZE_4K]; //ro
-extern __attribute__(( aligned(8) )) u64 __xmhfhic_x86vmx_cpuidtable[MAX_X86_APIC_ID]; //ro
+//extern __attribute__(( aligned(8) )) u64 __xmhfhic_x86vmx_cpuidtable[MAX_X86_APIC_ID]; //ro
+extern __attribute__(( aligned(8) )) u32 __xmhfhic_x86vmx_cpuidtable[MAX_X86_APIC_ID]; //ro
 
 //extern u64  __xmhfhic_exceptionstubs[]; //ro
 extern u32  __xmhfhic_exceptionstubs[]; //ro
